* Float coordinates implemented
* Item pickup improved
* Bug with Character selection in register window fixed
* Changing coordinates in time algorithm changed (a bit more accurate). Also Character's speed implemented
* CharacterMoved bug temporary fixed (setPosition problem)
* Update method improvements (not tested/completed yet)
+ GameData::Get[CurrentMapObject] methods
* IGUIChatBox flexibility

+++ RPGatorDll/CurrentMapObject.h
@@ -6,8 +6,8 @@ class __declspec(dllexport) CurrentMapObject :
 	public CurrentGameObject<T>
 {
 public:
-	int x; //in "our" coordinates, not in pixels
-	int y; //in "our" coordinates, not in pixels
+	double x; //in "our" coordinates, not in pixels
+	double y; //in "our" coordinates, not in pixels
 	Location* currentLocation; //Cross-link. Client does not use it. It uses Universe->currentLocation
 	ISceneNode* node;
 	ISceneNode* title;
